pyproject.toml
[project]
name = "rag-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.13.0",
    "altair<5",
    "datasets>=4.8.4",
    "dill>=0.4.1",
    "dotenv>=0.9.9",
    "edgartools>=5.26.0",
    "faiss-cpu>=1.13.2",
    "numpy>=2.4.3",
    "openai>=2.29.0",
    "pandas>=3.0.1",
    "peft>=0.18.1",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.3.0",
    "standard-imghdr>=3.13.0",
    "streamlit>=1.19.0",
    "tabulate>=0.10.0",
    "torch>=2.11.0",
    "tqdm>=4.67.3",
]
